Set up without restriction on howm many grid parts
And when typed 100, it is stretched to the right for looooong.
Is the flex column adding lenght to the screen? Don't even know if
that's the right wording. Basically the screen was set to max width
of 960px, but when typed 100, I got 100 grid bits, but they were added
to the side, just like flexbox adds int top to bottom column setup.
And the max-width doesn't work...